🐛 fix(Utils.java): change PublicWalletAddress class to use ByteBuf instead of byte[] for address field ✨ feat(Utils.java): add equals() method to PublicWalletAddress and TransactionSignature classes for better comparison 🔧 refactor(Utils.java): refactor walletAddressToString() and stringToWalletAddress() methods to use ByteBuf instead of byte[] for better performance and readability
The PublicWalletAddress class has been modified to use ByteBuf instead of byte[] for the address field. This change improves performance and allows for better manipulation of the address data. Additionally, the equals() method has been added to the PublicWalletAddress and TransactionSignature classes to provide a more reliable way to compare instances of these classes. Lastly, the walletAddressToString() and stringToWalletAddress() methods have been refactored to use ByteBuf instead of byte[] for better performance and readability.