🔧 chore(build.gradle): add CQEngine library as a dependency 🔧 chore(ActiveJSerializer.java): add ActiveJSerializer class for serializing and deserializing objects using ActiveJ The build.gradle file has been modified to include the CQEngine library as a dependency. This allows the application to use the CQEngine library for its functionality.
A new file, ActiveJSerializer.java, has been added. This file contains the ActiveJSerializer class, which is used for serializing and deserializing objects using ActiveJ. The ActiveJSerializer class provides methods for serializing and deserializing objects, as well as performing sanity tests on the objects to check if they can be serialized and deserialized successfully.

The changes were made to improve the functionality of the application by adding support for the CQEngine library and providing a convenient way to serialize and deserialize objects using ActiveJ.