🐛 fix(Transaction.java): update 'from' method in TransactionImpl builder to handle transaction fee 🐛 fix(TransactionImpl.java): set default value for signingKey to PublicKey.empty() to avoid null pointer exception
The 'from' method in the TransactionImpl builder has been updated to handle the case where the transactionDto has a transaction fee. If the transactionDto has a transaction fee, the 'from' method will set the 'from' field to PublicAddress.empty(), otherwise it will set it to PublicAddress.of(transactionDto.signingKey).

In the TransactionImpl class, the signingKey field has been set a default value of PublicKey.empty() to avoid a null pointer exception when comparing signing keys.